+++ src/main/java/workcenter/entidades/ActividadDiaria.java
@@ -6,25 +6,11 @@
 
 package workcenter.entidades;
 
-import java.io.Serializable;
-import java.util.Date;
-import javax.persistence.Basic;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.OneToOne;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
+import javax.persistence.*;
 import javax.validation.constraints.Size;
 import javax.xml.bind.annotation.XmlRootElement;
+import java.io.Serializable;
+import java.util.Date;
 
 /**
  *
@@ -56,7 +42,7 @@ public class ActividadDiaria implements Serializable {
     @Column(name = "id")
     private Integer id;
     @Column(name = "id_usuario")
-    private Integer idUsuario;
+    private Long idUsuario;
     @Column(name = "hora")
     private Integer hora;
     @JoinColumn(name = "id_tipo_actividad", referencedColumnName = "id")
@@ -88,11 +74,11 @@ public class ActividadDiaria implements Serializable {
         this.id = id;
     }
 
-    public Integer getIdUsuario() {
+    public Long getIdUsuario() {
         return idUsuario;
     }
 
-    public void setIdUsuario(Integer idUsuario) {
+    public void setIdUsuario(Long idUsuario) {
         this.idUsuario = idUsuario;
     }
