+++ src/main/java/workcenter/entidades/Cliente.java
@@ -6,17 +6,11 @@
 
 package workcenter.entidades;
 
-import java.io.Serializable;
-import javax.persistence.Basic;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.Table;
+import javax.persistence.*;
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 import javax.xml.bind.annotation.XmlRootElement;
+import java.io.Serializable;
 
 /**
  *
@@ -45,7 +39,7 @@ public class Cliente implements Serializable {
     @Column(name = "nombre")
     private String nombre;
     @Column(name = "rut")
-    private Integer rut;
+    private Long rut;
     @Column(name = "digitoverificador")
     private Character digitoverificador;
     @Size(max = 30)
@@ -81,11 +75,11 @@ public class Cliente implements Serializable {
         this.nombre = nombre;
     }
 
-    public Integer getRut() {
+    public Long getRut() {
         return rut;
     }
 
-    public void setRut(Integer rut) {
+    public void setRut(Long rut) {
         this.rut = rut;
     }
