+++ src/main/java/workcenter/entidades/Empresa.java
@@ -1,17 +1,9 @@
 package workcenter.entidades;
 
-import java.io.Serializable;
-import javax.persistence.Basic;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.Table;
+import javax.persistence.*;
 import javax.validation.constraints.Size;
 import javax.xml.bind.annotation.XmlRootElement;
+import java.io.Serializable;
 
 /**
  *
@@ -40,7 +32,7 @@ public class Empresa implements Serializable {
     @Column(name = "nombre")
     private String nombre;
     @Column(name = "rut")
-    private Integer rut;
+    private Long rut;
     @Column(name = "digitoverificador")
     private Character digitoverificador;
     @Size(max = 30)
@@ -79,11 +71,11 @@ public class Empresa implements Serializable {
         this.nombre = nombre;
     }
 
-    public Integer getRut() {
+    public Long getRut() {
         return rut;
     }
 
-    public void setRut(Integer rut) {
+    public void setRut(Long rut) {
         this.rut = rut;
     }
