+++ src/main/java/workcenter/entidades/TrazabilidadOt.java
@@ -5,24 +5,11 @@
  */
 package workcenter.entidades;
 
+import javax.persistence.*;
+import javax.xml.bind.annotation.XmlRootElement;
 import java.io.Serializable;
 import java.util.Date;
 import java.util.Objects;
-import javax.persistence.Basic;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
-import javax.persistence.NamedQueries;
-import javax.persistence.NamedQuery;
-import javax.persistence.Table;
-import javax.persistence.Temporal;
-import javax.persistence.TemporalType;
-import javax.xml.bind.annotation.XmlRootElement;
 
 /**
  *
@@ -47,7 +34,7 @@ public class TrazabilidadOt implements Serializable, Comparable<TrazabilidadOt> 
     @Column(name = "estado_id")
     private Integer estadoId;
     @Column(name = "autor")
-    private Integer autor;
+    private Long autor;
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "ejecutor", referencedColumnName = "rut")
     private Personal ejecutor;
@@ -86,11 +73,11 @@ public class TrazabilidadOt implements Serializable, Comparable<TrazabilidadOt> 
         this.estadoId = estadoId;
     }
 
-    public Integer getAutor() {
+    public Long getAutor() {
         return autor;
     }
 
-    public void setAutor(Integer autor) {
+    public void setAutor(Long autor) {
         this.autor = autor;
     }
